added private endpoints, modified auth

+++ backend/content-hub/api/v1/articles.py
@@ -1,9 +1,10 @@
 from typing import Annotated, Sequence
 
-from fastapi import APIRouter, Depends, status, Query
+from fastapi import APIRouter, Depends, status, Query, HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from core.models import Article, db_helper
+from api.auth.auth_config import get_current_auth_user
+from core.models import Article, db_helper, User
 from core.schemas.article import (
     ArticleCreateSchema,
     ArticleReadSchema,
@@ -19,7 +20,13 @@ router = APIRouter(tags=["Articles"])
 async def create_article_endpoint(
     session: Annotated[AsyncSession, Depends(db_helper.session_getter)],
     article_create: ArticleCreateSchema,
+    current_user: User = Depends(get_current_auth_user)
 ) -> Article:
+    if current_user.id != article_create.user_id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="No access."
+        )
     article = await articles_crud.create_article(
         session=session,
         article_create=article_create,
@@ -56,7 +63,15 @@ async def update_article_endpoint(
     article_id: int,
     article_update: ArticleUpdateSchema,
     session: Annotated[AsyncSession, Depends(db_helper.session_getter)],
+    current_user: User = Depends(get_current_auth_user)
 ) -> Article:
+    article1 = await articles_crud.get_article(session, article_id)
+    if article1.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="No access."
+        )
+
     article = await articles_crud.update_article(
         session=session,
         article_id=article_id,
@@ -69,7 +84,15 @@ async def update_article_endpoint(
 async def delete_article_endpoint(
     article_id: int,
     session: Annotated[AsyncSession, Depends(db_helper.session_getter)],
+    current_user: User = Depends(get_current_auth_user)
 ) -> None:
+    article1 = await articles_crud.get_article(session, article_id)
+    if article1.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="No access."
+        )
+
     await articles_crud.delete_article(
         session=session,
         article_id=article_id,
